requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
httpx==0.28.1
google-auth==2.40.1
pydantic==2.11.3
pydantic-settings==2.8.1
pytest==8.3.5
pytest-asyncio==0.26.0
